green - SetY_double

+++ RayTracingApp/Engine/Vector.cs
@@ -9,6 +9,6 @@ namespace Engine
     public class Vector
     {
         public double X {  get; set; }
-        public int Y { get; set; }
+        public double Y { get; set; }
     }
 }
